Use coproc only for communicating the process substitution exit code
Turns out that processes started via coproc do not inherit the PTY,
which means commands which prompt on TTY (pass for example) would fail
to execute in macros correctly.

The previous implementation possibly started two coprocesses at the same
time, which is not officially supported.

This switches to using coproc only for communicating the exit code of
the rendered_script() process substitution.