Use coproc only for communicating the process substitution exit code
Turns out that processes started via coproc do not inherit the PTY,
which means commands which prompt on TTY (pass for example) would fail
to execute in macros correctly.

The previous implementation also started one coprocess inside another,
not sure if that's a supported configuration.

Switch to using coproc only for communicating the exit code of
the rendered_script() process substitution.